requirements.txt
absl-py==1.2.0
Adafruit-Blinka==6.20.1
adafruit-circuitpython-neopixel==6.2.3
adafruit-circuitpython-pixelbuf==1.1.1
Adafruit-PlatformDetect==3.19.4
Adafruit-PureIO==1.1.9
aiofiles==0.6.0
aiohttp==3.8.1
aiosignal==1.2.0
aiosqlite==0.17.0
airdrop-ng==1.1
airgraph-ng==1.1
anyio==3.5.0
appdirs==1.4.4
application-utility==1.3.2
asgiref==3.5.0
astroid==2.14.2
asttokens==2.2.1
astunparse==1.6.3
async-generator==1.10
async-timeout==4.0.2
asyncio-extras==1.3.2
asyncpraw==7.5.0
asyncprawcore==2.3.0
attrs==22.2.0
autocommand==2.2.2
autopep8==1.6.0
beautifulsoup4==4.12.0
bitstring==3.1.9
black==21.12b0
bottle==0.12.19
Brotli==1.0.9
bs4==0.0.1
btrfsutil==6.1.3
build==0.10.0
CacheControl==0.12.11
cachetools==5.2.0
certifi==2022.12.7
cffi==1.15.1
chardet==5.1.0
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
commonmark==0.9.1
conan==1.44.0
configparser==5.2.0
contourpy==1.0.7
crayons==0.4.0
cryptography==39.0.1
cssselect==1.1.0
cupshelpers==1.0
cycler==0.11.0
Cython==0.29.33
Deprecated==1.2.13
dill==0.3.6
distlib==0.3.6
distro==1.8.0
docopt==0.6.2
docutils==0.19
ecdsa==0.18.0
esptool==4.4
fake-useragent==0.1.11
fastapi==0.89.1
fasteners==0.16.3
fastjsonschema==2.16.2
ffmpeg==1.4
flatbuffers==1.12
fonttools==4.33.3
frozenlist==1.3.0
future==0.18.2
gast==0.4.0
gevent==21.12.0
gitdb==4.0.10
GitPython==3.1.29
google-auth==2.10.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grapejuice @ file:///home/clay/.cache/yay/grapejuice-git/src/grapejuice
Graphine==0.0
greenlet==1.1.2
grequests==0.6.0
grpcio==1.47.0
gTTS==2.2.3
guake==3.9.1.dev0
h11==0.13.0
h5py==3.7.0
idna==3.4
importlib-metadata==5.0.0
inflect==6.0.2
iniconfig==1.1.1
installer==0.6.0
isort==5.12.0
jaconv==0.3.4
jaraco.context==4.3.0
jaraco.functools==3.5.2
jaraco.text==3.11.1
jedi==0.18.2
Jinja2==3.0.3
keras==2.9.0
Keras-Preprocessing==1.1.2
keyutils==0.6
kiwisolver==1.4.2
lazy-object-proxy==1.9.0
lensfun==0.3.3
libclang==14.0.6
libfdt==1.6.1
libvirt-python==9.0.0
lit==15.0.7.dev0
lxml==4.7.1
Mako==1.2.4
manjaro-sdk==0.1
Markdown==3.4.1
markdown-it-py==2.1.0
MarkupSafe==2.1.1
matplotlib==3.5.2
mccabe==0.7.0
mdurl==0.1.2
mido==1.2.10
more-itertools==9.0.0
MouseInfo==0.1.3
msgpack==1.0.4
multidict==6.0.2
mutagen==1.46.0
mypy==1.0.0
mypy-extensions==0.4.3
netsnmp-python==1.0a1
nftables==0.1
node-semver==0.6.1
npyscreen==4.10.5
nspektr==0.4.0
numpy==1.24.1
oauthlib==3.2.0
openai==0.27.1
opencv-python==4.5.5.62
opt-einsum==3.3.0
ordered-set==4.1.0
oscpy==0.6.0
outcome==1.1.0
packaging==21.3
pacman-mirrors==4.23.2
pandas==1.3.5
parse==1.19.0
parso==0.8.3
patch-ng==1.17.4
pathspec==0.9.0
pbr==5.11.1
pep517==0.13.0
Pillow==9.4.0
Pivy==0.6.8
platformdirs==2.6.2
pluggy==1.0.0
pluginbase==1.0.1
ply==3.11
praw==7.5.0
prawcore==2.3.0
progressbar2==3.55.0
protobuf==3.19.4
psutil==5.9.4
ptyprocess==0.7.0
py==1.11.0
pyaes==1.6.1
PyAutoGUI==0.9.53
pybind11==2.10.3
pycairo==1.23.0
pycodestyle==2.8.0
pycparser==2.21
pycryptodomex==3.17
pycups==2.0.1
pycurl==7.45.2
pydantic==1.10.7
pyee==8.2.2
pyfiglet==0.8.post1
pyftdi==0.53.3
PyGetWindow==0.0.9
Pygments==2.14.0
PyGObject==3.42.2
PyJWT==1.7.1
pykakasi==2.2.1
pylint==2.16.2
PyMsgBox==1.0.9
pyOpenSSL==23.0.0
pyparsing==3.0.9
pyperclip==1.8.2
pyppeteer==1.0.2
pyproject_hooks==1.0.0
PyQt5==5.15.9
PyQt5-sip==12.11.1
PyQt6==6.3.1
PyQt6-Qt6==6.3.2
PyQt6-sip==13.4.0
pyquery==1.4.3
PyRect==0.1.4
Pyro4==4.82
PyScreeze==0.1.28
pyserial==3.5
PySide2==5.15.2.1
PySide6==6.4.2
pysmbc==1.0.23
pytest==7.1.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-slugify==8.0.1
python-utils==2.7.1
python3-xlib==0.15
pytrends==4.7.3
pytube==12.1.2
pytweening==1.0.4
pytz==2021.3
pyusb==1.2.1
PyYAML==6.0
QTermWidget==1.2.0
rapidfuzz==2.13.7
redis==4.5.3
reedsolo==1.7.0
reportlab==3.6.12
requests==2.28.2
requests-html==0.10.0
requests-oauthlib==1.3.1
resolvelib==0.9.0
rich==13.3.1
rsa==4.9
schedule==1.1.0
selenium==4.1.0
Send2Trash==1.8.0
serpent==1.41
shiboken2==5.15.2.1
shiboken6==6.4.2
shiboken6-generator==6.4.2
shtab==1.5.8
six==1.16.0
smmap==5.0.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.1
sox==1.4.1
spotdl==4.1.3
spotipy==2.22.1
SQLAlchemy==1.4.35
starlette==0.22.0
syncedlyrics==0.2.2
systemd-python==235
TBB==0.2
team==1.0
tenacity==8.2.1
tensorboard==2.9.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow-estimator==2.9.0
tensorflow-gpu==2.9.1
tensorflow-io-gcs-filesystem==0.26.0
termcolor==1.1.0
text-unidecode==1.3
thonny==4.0.2
tldr==3.1.0
toml==0.10.2
tomli==2.0.1
tomlkit==0.11.6
tqdm==4.65.0
trio==0.19.0
trio-websocket==0.9.2
trove-classifiers==2023.2.9
typing_extensions==4.5.0
uc-micro-py==1.0.1
udiskie==2.4.2
Unidecode==1.3.6
update-checker==0.18.0
urllib3==1.26.15
urwid==2.1.2
uvicorn==0.20.0
validate-pyproject==0.12.1
w3lib==1.22.0
webdriver-manager==3.5.2
webencodings==0.5.1
websocket-client==1.2.3
websockets==10.2
Werkzeug==2.2.2
wikipedia==1.4.0
wrapt==1.14.1
wsproto==1.0.0
yarl==1.7.2
youtube-dl==2021.12.17
yt-dlp==2023.3.4
ytmusicapi==0.24.1
zipp==3.13.0
zope.event==4.5.0
